requirements.txt
# hydra-core==1.1.1
# chess==1.6.1
# easy-to-hard-data==1.0.0
jax[cuda]==0.3.10+cuda11.cudnn82
equinox==0.5.0
tqdm==4.64.0
optax==0.1.2
matplotlib==3.5.2
wandb
hydra-core==1.1.2
hydra-colorlog==1.1.0